Completed project
As time ran out, I am committing my project.  I was able to create a
main page, just in case I would want to add more to this project.  I
also created the user cards, that will show an avatar to those who don't
have one in their profile.  The title and emails are only shown if the
employee has one in their file.  I also used a useState for the Follow
button.